+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "engine",
     "prelude",
+    "sdk"
 ]
 exclude = [
     "etc/state-migration-test"
