+++ engine/src/types.rs
@@ -4,8 +4,6 @@ use prelude::{self, str, Address, String, ToString, Vec, H256, U256};
 #[cfg(not(feature = "contract"))]
 use sha3::{Digest, Keccak256};
 
-use crate::sdk;
-
 pub type AccountId = String;
 pub type Balance = u128;
 pub type RawAddress = [u8; 20];
@@ -197,23 +195,6 @@ pub fn bytes_to_hex(v: &[u8]) -> String {
     result
 }
 
-#[cfg(feature = "contract")]
-#[inline]
-pub fn keccak(data: &[u8]) -> H256 {
-    sdk::keccak(data)
-}
-
-#[cfg(not(feature = "contract"))]
-#[inline]
-pub fn keccak(data: &[u8]) -> H256 {
-    H256::from_slice(Keccak256::digest(data).as_slice())
-}
-
-#[allow(dead_code)]
-pub fn near_account_to_evm_address(addr: &[u8]) -> Address {
-    Address::from_slice(&keccak(addr)[12..])
-}
-
 #[derive(Default)]
 pub struct Stack<T> {
     stack: Vec<T>,
@@ -256,185 +237,4 @@ pub fn str_from_slice(inp: &[u8]) -> &str {
 #[cfg(feature = "contract")]
 pub trait ExpectUtf8<T> {
     fn expect_utf8(self, message: &[u8]) -> T;
-}
-
-#[cfg(feature = "contract")]
-impl<T> ExpectUtf8<T> for Option<T> {
-    fn expect_utf8(self, message: &[u8]) -> T {
-        match self {
-            Some(t) => t,
-            None => sdk::panic_utf8(message),
-        }
-    }
-}
-
-#[cfg(feature = "contract")]
-impl<T, E> ExpectUtf8<T> for core::result::Result<T, E> {
-    fn expect_utf8(self, message: &[u8]) -> T {
-        match self {
-            Ok(t) => t,
-            Err(_) => sdk::panic_utf8(message),
-        }
-    }
-}
-
-pub trait SdkExpect<T> {
-    fn sdk_expect(self, msg: &str) -> T;
-}
-
-impl<T> SdkExpect<T> for Option<T> {
-    fn sdk_expect(self, msg: &str) -> T {
-        match self {
-            Some(t) => t,
-            None => sdk::panic_utf8(msg.as_ref()),
-        }
-    }
-}
-
-impl<T, E> SdkExpect<T> for core::result::Result<T, E> {
-    fn sdk_expect(self, msg: &str) -> T {
-        match self {
-            Ok(t) => t,
-            Err(_) => sdk::panic_utf8(msg.as_ref()),
-        }
-    }
-}
-
-pub trait SdkUnwrap<T> {
-    fn sdk_unwrap(self) -> T;
-}
-
-impl<T> SdkUnwrap<T> for Option<T> {
-    fn sdk_unwrap(self) -> T {
-        match self {
-            Some(t) => t,
-            None => sdk::panic_utf8("ERR_UNWRAP".as_bytes()),
-        }
-    }
-}
-
-impl<T, E: AsRef<[u8]>> SdkUnwrap<T> for core::result::Result<T, E> {
-    fn sdk_unwrap(self) -> T {
-        match self {
-            Ok(t) => t,
-            Err(e) => sdk::panic_utf8(e.as_ref()),
-        }
-    }
-}
-
-pub(crate) trait SdkProcess<T> {
-    fn sdk_process(self);
-}
-
-impl<T: AsRef<[u8]>, E: AsRef<[u8]>> SdkProcess<T> for Result<T, E> {
-    fn sdk_process(self) {
-        match self {
-            Ok(r) => sdk::return_output(r.as_ref()),
-            Err(e) => sdk::panic_utf8(e.as_ref()),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_hex() {
-        assert_eq!(
-            bytes_to_hex(&[0u8, 1u8, 255u8, 16u8]),
-            "0001ff10".to_string()
-        );
-    }
-
-    /// Build view of the stack. Intervals between None values are scopes.
-    fn view_stack(stack: &Stack<i32>) -> Vec<Option<i32>> {
-        let mut res = vec![];
-        let mut pnt = 0;
-
-        for &pos in stack.boundaries.iter() {
-            while pnt < pos {
-                res.push(Some(stack.stack[pnt]));
-                pnt += 1;
-            }
-            res.push(None);
-        }
-
-        while pnt < stack.stack.len() {
-            res.push(Some(stack.stack[pnt]));
-            pnt += 1;
-        }
-
-        res
-    }
-
-    fn check_stack(stack: &Stack<i32>, expected: Vec<Option<i32>>) {
-        if let Some(&last) = stack.boundaries.last() {
-            assert!(last <= stack.stack.len());
-        }
-        assert_eq!(view_stack(stack), expected);
-    }
-
-    #[test]
-    fn test_stack() {
-        let mut stack = Stack::new(); // [ $ ]
-        check_stack(&stack, vec![None]);
-
-        stack.push(1); // [ $, 1]
-        check_stack(&stack, vec![None, Some(1)]);
-        stack.push(2); // [ $, 1, 2 ]
-        check_stack(&stack, vec![None, Some(1), Some(2)]);
-        stack.enter(); // [$, 1, 2, $]
-        check_stack(&stack, vec![None, Some(1), Some(2), None]);
-        stack.push(3); // [$, 1, 2, $, 3]
-        check_stack(&stack, vec![None, Some(1), Some(2), None, Some(3)]);
-        stack.discard(); // [$, 1, 2]
-        check_stack(&stack, vec![None, Some(1), Some(2)]);
-        stack.enter();
-        check_stack(&stack, vec![None, Some(1), Some(2), None]);
-        stack.push(4); // [$, 1, 2, $, 4]
-        check_stack(&stack, vec![None, Some(1), Some(2), None, Some(4)]);
-        stack.enter(); // [$, 1, 2, $, 4, $]
-        check_stack(&stack, vec![None, Some(1), Some(2), None, Some(4), None]);
-        stack.push(5); // [$, 1, 2, $, 4, $, 5]
-        check_stack(
-            &stack,
-            vec![None, Some(1), Some(2), None, Some(4), None, Some(5)],
-        );
-        stack.commit(); // [$, 1, 2, $, 4, 5]
-        check_stack(&stack, vec![None, Some(1), Some(2), None, Some(4), Some(5)]);
-        stack.discard(); // [$, 1, 2]
-        check_stack(&stack, vec![None, Some(1), Some(2)]);
-        stack.push(6); // [$, 1, 2, 6]
-        check_stack(&stack, vec![None, Some(1), Some(2), Some(6)]);
-        stack.enter(); // [$, 1, 2, 6, $]
-        check_stack(&stack, vec![None, Some(1), Some(2), Some(6), None]);
-        stack.enter(); // [$, 1, 2, 6, $, $]
-        check_stack(&stack, vec![None, Some(1), Some(2), Some(6), None, None]);
-        stack.enter(); // [$, 1, 2, 6, $, $, $]
-        check_stack(
-            &stack,
-            vec![None, Some(1), Some(2), Some(6), None, None, None],
-        );
-        stack.commit(); // [$, 1, 2, 6, $, $]
-        check_stack(&stack, vec![None, Some(1), Some(2), Some(6), None, None]);
-        stack.discard(); // [$, 1, 2, 6, $]
-        check_stack(&stack, vec![None, Some(1), Some(2), Some(6), None]);
-        stack.push(7); // [$, 1, 2, 6, $, 7]
-
-        assert_eq!(stack.into_vec(), vec![1, 2, 6, 7]);
-    }
-
-    #[test]
-    fn test_wei_from_u64() {
-        let x: u64 = rand::random();
-        assert_eq!(Wei::new_u64(x).raw().as_u64(), x);
-    }
-
-    #[test]
-    fn test_wei_from_eth() {
-        let eth_amount: u64 = rand::random();
-        let wei_amount = U256::from(eth_amount) * U256::from(10).pow(18.into());
-        assert_eq!(Wei::from_eth(eth_amount.into()), Some(Wei::new(wei_amount)));
-    }
 }
